Allow `tag` to be used as `_type` property.
Take the last tag component and set it as the `_type` property. e.g. `foo.bar.baz` gets `_type: "baz"`. For something like the following use case:

```javascript
fluent.configure('foo');
fluent.emit('bar', { ... });
```

It is quite handy to capture the `my-event` as the `_type` property. I think this is a common enough behavior to warrant a configuration option for. It's also impossible to configure with any of the existing options; using `tag_name "#{tag}"` fails since the evaluation happens when the config file is parsed, not when the message is processed.

Specs added for this new behavior.