Addresses invalid timestamp data causing a wasteful retry loop
The main issue is that there is no checking for parse errors when
trying to parse a passed in `@timestamp` of `time_key` value.  If
`nil` is passed in by accident, you get a `TypeError` as the code
tries to convert `nil` to a string.  If a garbage string with no
numbers is passed in, `Time.parse` throws an `ArgumentError`. Both
of these non-recoverable errors are bad because it sends the plugin
into an infinite fluentd plugin retry loop.  Since retrying is not
going to correct the bad data, the best solution is to proceed with
the record's time as the timestamp and additionally log a warning
message about the parse issue and the offending log data.