Add support for `target_index_key` config.
The purpose of this is to allow a (custom) filter to tell this plugin the index name to write to, saving the expensive `Time.parse` operation when it can be pre-determined.

If you've set this config option but not provided a value it will fallback to the previous behaviour.

This is one of the PRs to come out of uken/fluent-plugin-elasticsearch#152 "Time.parse is slow - allow other methods of finding index name"