Adds `remove_keys_on_update` feature
It is useful especially when using upsert to skip some keys
from the update document. This feature enables that.

We are using fluentd to load and update some datasets in es.
Another application then can update certain fields to mark the
status of a particular record. So we wan't the data to be loaded
with some initial values for these fields, and then on future
updates, not change these fields.