Allow logstash prefix via key in record
Today the prefix used for the index name of the target index to send
records is fixed at configuration time when not using the dynamic
plug-in.

This commit provides the user with the ability to configure what key a
record to use to pull the index prefix from when constructing the
Elasticsearch bulk request.  This almost entirely obviates the need for
the dynamic plug-in version as a result.