WIP: Retry on certain errors from Elasticsearch
This is a work-in-progress: we need unit tests to exercise the error
handling code paths, and we to ensure the caller can configure an
alternate output plugin to handle errors if they so choose. For now, we
just always raise some kind of exception if we get a response from
Elasticsearch.

We specifically handle two retry conditions, bulk index queue overflow,
and java out of memory. In both cases, we should retry until the chunk
is indexed, and we raise exceptions to allow the generic output methods
to handle the retry.

We also detect an Elasticsearch version mismatch when the returned data
structure does not match what we expect, and we accommodate other error
conditions generically.

Note we only perform these changes currently for the "static" version of the
plugin.