auth: Fix missing auth tokens after reloading connections
This patch fixes the auth failure issue reported by several users.

Essentially, elasticsearch-ruby treats the `hosts` info as something
ephemeral and try to reconstruct it periodically using the latest
information fetched from the cluster.

Since this, as to the settings which should be persistent across the
refresh (e.g. an authentication token), clients are supposed to pass
them through `options` parameter, not as a part of `hosts`. Otherwise,
these bits may be lost in the process of refreshing connections.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>