handle case where stats not processed in order; add testing
The code did not return/raise consistent errors.  This is because
it is iterating the stats Hash which could return keys in any
order, resulting in non-deterministic behavior.
The fix is to first look for known success and error conditions
and return those first, then return cases where the errors
are unknown.

This also adds tests for the above cases.