Add an option `include_index_in_url` to allow URL-based conrtrols
This option tells Fluentd to manifest the name of the target index in
the request URL (rather than in the request body). For example:

    POST /logstash-20180501/_bulk

And without this option, Fluentd just uses the general endpoint:

    POST /_bulk

We need this option in order to enforce URL-based access controls on
Elasticsearch endpoints. For details, read the following article:

https://www.elastic.co/guide/en/elasticsearch/reference/1.5/url-access-control.html