Implement the `include_index_in_url` option for out_elasticsearch
Now, you can use this option with the normal elasticsearch plugin.

The nasty thing here is that we cannot assume all events in a single
chunk share the same index name even under a non-dynamic plugin. Why?
Because Elasticsearch's index names usually contain a timestamp (e.g.
'logstash-2018.01.01') and these timestamps are generated from each
record's time field.

This is why we had to convert `bulk_message` into a hash, and do all
the work to get the other methods working fine with it.