Fix ILM index template order calculation when index_separator is not "-"
When creating index templates the `order`/`priority` was calculated using the number of
`-` characters in the target index name.  This was seemingly intended to give the effect of
a "longest prefix match" for a template name to an index.

For example, if you have an indexes `logs-http` and `logs-http-access` then the these templates
are generated with patterns `logs-http-*` and `logs-http-access-*`.  An index named
`logs-http-access-default-00001` would match both of these patterns, so we need a higher
`order`/`priority` on the template for `logs-http-access`.

This worked correctly if `index_separator` was set to `-` because the separator used before the wildcard
was `-` so that was the separator to count for the priority.  However, if `index-separator` is set to
another character, such as `.`, then we need to count occurrences of that separator instead.

The updated example might be `logs.http.*` and `logs.http.access.*`.  The second template should have a
higher priority than the first, but it previously wouldn't because it didn't have any `-` characters
in it.