Avoid long worrying log line at info level
Currently there is an info level log line when checking if a datastream exists, but it includes the entire error even though it is caught and expected.

This results in long worrying looking log messages which are in fact not a problem (once you notice that they are info level) e.g. like this:
```
[info]: #0 Specified data stream does not exist. Will be created: <[404] {"error":{"root_cause":[{"type":"index_not_found_exception","reason":"no such index [my.datastream.name]","resource.type":"index_or_alias","resource.id":"my.datastream.name","index_uuid":"_na_","index":"my.datastream.name"}],"type":"index_not_found_exception","reason":"no such index [my.datastream.name]","resource.type":"index_or_alias","resource.id":"my.datastream.name","index_uuid":"_na_","index":"my.datastream.name"},"status":404}>
```
so simplify it to instead just log the name of the datastream.

We know what the error is going to be given it is being caught explicitly.